+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # [next]
 
 - add params to replace in getString
+- add `T getCube<T extends Cube>()`
 
 # 0.3.1